requirements.txt
pytest==7.4.0
requests==2.31.0
jsonschema==4.19.0
pytest-html==3.2.0
allure-pytest==2.11.0
pytest-xdist==3.3.1
coverage==7.2.7
python-dotenv==1.0.0
pytest-rerunfailures==10.3